xml/parsing/manakai/eldecls-perefs-2.dat

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ELEMENT b (%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;14;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "  ">
<!ELEMENT b (%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;14;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ELEMENT b ( %empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;15;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ELEMENT b (a%empty;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;15;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ELEMENT b (a %empty;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;16;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ELEMENT b (c|%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;16;xml:dtd:cm:empty entity;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (c | a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ENTITY % foo "%empty;">
<!ELEMENT b (%foo;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]4;14;xml:dtd:cm:empty entity;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "|">
<!ELEMENT b (b%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]2;19;xml:dtd:cm:entity begins with connector;;;s
[1]3;15;xml:dtd:cm:entity ends with connector;;;s
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b | a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "  |
">
<!ELEMENT b (b%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]2;21;xml:dtd:cm:entity begins with connector;;;s
[1]4;15;xml:dtd:cm:entity ends with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b | a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "  ,
">
<!ELEMENT b (b%empty;a)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]2;21;xml:dtd:cm:entity begins with connector;;;s
[1]4;15;xml:dtd:cm:entity ends with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b, a)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "  | d
">
<!ELEMENT b (b%empty;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]2;21;xml:dtd:cm:entity begins with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b | d)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "  | (d)
">
<!ELEMENT b (b%empty;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]2;21;xml:dtd:cm:entity begins with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b | (d))>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty SYSTEM "http://f/">
<!ELEMENT b (b%empty;)>
#resource http://f/
  | (d)
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;3;xml:dtd:cm:entity begins with connector;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (b | (d))>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "   d|
">
<!ELEMENT b (%empty;b)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]4;14;xml:dtd:cm:entity ends with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (d | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "   (d)|
">
<!ELEMENT b (%empty;b)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]4;14;xml:dtd:cm:entity ends with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b ((d) | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % e "f">
<!ENTITY % empty "   (%e;)|
">
<!ELEMENT b (%empty;b)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]5;1;xml:dtd:ext decl;;;w
[1]5;14;xml:dtd:cm:entity ends with connector;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b ((f) | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % e SYSTEM "http://f/">
<!ENTITY % empty "   (%e;)|
">
<!ELEMENT b (%empty;b)>
#resource http://f/
<?xml encoding="utf-8"?>f
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]5;1;xml:dtd:ext decl;;;w
[1]5;14;xml:dtd:cm:entity ends with connector;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b ((f) | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ENTITY % foo "%empty; | b">
<!ELEMENT b (a%foo;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]3;25;xml:dtd:cm:entity begins with connector;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ENTITY % foo "&#x25;empty; | b">
<!ELEMENT b (a%foo;)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]3;17;xml:dtd:cm:empty entity;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % empty "">
<!ENTITY % foo "a|&#x25;empty;">
<!ELEMENT b (%foo;b)>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
[1]3;19;xml:dtd:cm:empty entity;;;s
[1]4;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a | b)>
| <{} a>

#data
<?xml version="1.0"?>
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<?xml encoding="utf-8"?>
<!ENTITY % e "(a)">
<!ELEMENT b %e;>
#errors
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT b (a)>
| <{} a>
